Split common test functions into a new file
