+++ tests/stats_test.rs
@@ -1,30 +1,9 @@
-extern crate pihole_api;
 extern crate rmp;
 
-use std::collections::HashMap;
+mod common;
+
 use rmp::encode;
-
-/// Test an API endpoint by inputting test data and checking the response
-fn test_endpoint(endpoint: &str, ftl_command: &str, ftl_data: Vec<u8>, expected: &str) {
-    // Add the test data
-    let mut data = HashMap::new();
-    data.insert(ftl_command.to_owned(), ftl_data);
-
-    // Start the test client
-    let client = pihole_api::test(data);
-
-    // Get the response
-    let mut response = client.get(endpoint).dispatch();
-    let body = response.body_string();
-
-    // Check against expected output
-    assert!(body.is_some());
-    assert_eq!(expected, body.unwrap());
-}
-
-fn write_eom(data: &mut Vec<u8>) {
-    data.push(0xc1);
-}
+use common::{test_endpoint, write_eom};
 
 #[test]
 fn test_summary() {
